Using histograms to compute and plot the PDF and sigmas
Added functionality to use the histogram instead of a KDE in
plot_contours(). It does not actually draw contour lines yet
but plots a histogram instead of the contourf colors.

We can either plot the "sigmas" (equivalent to the current contourf
but based on a histogram instead of a KDE), or just plot the PDF
(i.e. histogram bin values).